Move diagnostic model to Lib
